requirements.txt
appdirs==1.4.0
astroid==1.4.9
autopep8==1.2.4
Django==1.10.1
django-cors-headers==2.0.2
django-filters==0.2.1
django-webpack-loader==0.3.3
djangorestframework==3.4.6
flake8==3.3.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
packaging==16.8
pep8==1.7.0
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.1.10
six==1.10.0
wrapt==1.10.8
